Move keep-related functions to separate file in chain packages
celo and ethereum both gain bonded_ecdsa_keep_handle.go files that group
all functions that currently live under the BondedECDSAKeep interface.
This is in preparation for modifying that interface to represent a
handle to a single keep, rather than embedding it in the chain.Handle
interface directly and having all functions parametrized by the keep ID.
